+++ app/About.tsx
@@ -35,14 +35,14 @@ export default function Home() {
         <Native.View style={styles.containerview}>
             <Native.ScrollView style={styles.mainview}>
                 <BetterText
-                    fontSize={25}
+                    fontSize={20}
                     fontWeight="Light"
                     onTap={() => Router.router.navigate("/Profile")}
                 >
                     {"<"} Go back
                 </BetterText>
                 <GapView height={20} />
-                <BetterText textAlign="normal" fontWeight="Bold" fontSize={40}>
+                <BetterText textAlign="normal" fontWeight="Bold" fontSize={30}>
                     About PersonaPlus
                 </BetterText>
                 <GapView height={20} />
@@ -52,11 +52,14 @@ export default function Home() {
                         header="What is PersonaPlus?"
                     >
                         <BetterText fontWeight="Regular" fontSize={15}>
-                            It&apos;s clear what it is. An app designed to help
+                            {/* It&apos;s clear what it is. An app designed to help
                             you be better. Currently it&apos;s on a very early
                             stage of development, so I can&apos;t say (yet)
                             it&apos;s a revolution. But it is, indeed, a piece
-                            of hard work designed to make your life better.
+                            of hard work designed to make your life better. */}
+                            It&apos;s clear: an app designed to help you give
+                            yourself more and have a better life through help
+                            and hard work.
                         </BetterText>
                     </Division>
                 </Section>
@@ -67,7 +70,7 @@ export default function Home() {
                         header="Who is PersonaPlus?"
                     >
                         <BetterText fontWeight="Regular" fontSize={15}>
-                            I&apos;m Zaka and I Do Stuff. I&apos;ve been since
+                            {/* I&apos;m Zaka and I Do Stuff. I&apos;ve been since
                             2023 working on this as my first ever React project.
                             Yep, this is my first time coding with the
                             technology that powers the app! I didn&apos;t want
@@ -78,32 +81,27 @@ export default function Home() {
                             Also, at the moment there aren&apos;t any
                             contributors yet, but remember this is a fully open
                             source app! You can help it grow at any time, and
-                            get credited here - as deserved!
+                            get credited here - as deserved! */}
+                            PersonaPlus is an independent app made by myself
+                            (Zaka). However, it is an open source project, you
+                            can contribute to it&apos;s code (and get your name
+                            here!)
                         </BetterText>
                     </Division>
                 </Section>
-                <GapView height={40} />
-                <BetterText
-                    textAlign="center"
-                    fontWeight="Bold"
-                    fontSize={40}
-                    textColor="#32FF80"
-                >
-                    Give yourself a plus!
-                </BetterText>
-                <GapView height={40} />
+                <GapView height={20} />
                 <Button
                     buttonText="License"
                     style="GOD"
                     action={() => Router.router.navigate("/License")}
                 />
-                <GapView height={15} />
+                <GapView height={10} />
                 <Button
                     buttonText="OSS used"
                     style="ACE"
                     action={() => Router.router.navigate("/OpenSourceCredits")}
                 />
-                <GapView height={15} />
+                <GapView height={10} />
                 <Button
                     buttonText="Privacy policy (external link)"
                     style="ACE"
@@ -113,6 +111,15 @@ export default function Home() {
                         )
                     }
                 />
+                <GapView height={20} />
+                <BetterText
+                    textAlign="center"
+                    fontWeight="Bold"
+                    fontSize={30}
+                    textColor="#32FF80"
+                >
+                    Give yourself a plus!
+                </BetterText>
             </Native.ScrollView>
         </Native.View>
     );